[0x05 - c - task 10] long replaced by int

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -14,7 +14,7 @@ int _atoi(char *s)
 {
 	int i, digit, len;
 	int nstart, order;
-	long int result;
+	int result;
 	int nminus;
 
 	len = _strlen(s);
